backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
groq
chromadb
pydantic
pydantic-settings
python-magic
PyPDF2
sentence-transformers
